create HasEmissions concern
